+++ src/api/product.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import API_URL from '@src/constants/apiUrlConst';
-import { IProduct } from "@src/interfaces/product";
 import { axiosInstance } from './instance';
+import { IProduct } from "@src/env";
 
 export const getProducts = async () => {
   const data = await axios.get('/mockData/product.json')
