--- src/interfaces/signup.ts
@@ -1,9 +1,0 @@
-export interface ISignup {
-  id: string
-  pw:string
-  name: string
-  nickname: string
-  phoneNumber: string
-  birthDate: string
-  hobby: string
-}
